simple fix on delete deck

+++ delete_deck.php
@@ -9,8 +9,11 @@ header('Content-Type: application/json');
 
 session_start();
 $user_id = 1;
-$data = json_decode(file_get_contents("php://input"), true);
-$id = $data['id'];
+if (!isset($_GET['id'])) {
+  echo json_encode(['error' => 'Missing deck ID']);
+  exit;
+}
+$id = (int)$_GET['id'];
 
 $conn->query("DELETE FROM deck_words WHERE deck_id = $id");
 $conn->query("DELETE FROM decks WHERE id = $id AND user_id = " . $user_id);
